requirements.txt
python-dotenv==1.0.1
html-table-parser-python3==0.3.1
pandas==2.2.1
google-cloud-core==2.4.1
google-cloud-storage==2.16.0
google-cloud-bigquery==3.20.1
google-api-core==2.18.0